requirements.txt
tweepy>=4.14
vaderSentiment>=3.3
python-dotenv>=1.0
